refactor(cardsection): переименовал CardSection,CatalogCardSection,убрал Carousel в PopularProducts
