+++ src/components/CardSection/CardSection.css
@@ -1,10 +1,8 @@
 .card-section {
-  box-sizing: border-box;
-}
-
-.card-section__header {
-  display: flex;
-  margin: 100px 0 40px 0;
+  margin: 0 auto;
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 32px;
 }
 
 .card-section__filter-header {
@@ -62,17 +60,6 @@
   line-height: 1.63;
   color: #605f5f;
   margin: 0 0 15px 0;
-}
-
-.card-section__title {
-  font-family: 'Didact-Gothic';
-  font-size: 40px;
-  font-weight: 400;
-  line-height: 1.3;
-  letter-spacing: -0.5px;
-  margin: 0;
-  padding-left: 40px;
-  color: #21211b;
 }
 
 .card-section__link {
