+++ src/components/CardSectionWithTitle/CardSectionWithTitle.css
@@ -1,0 +1,23 @@
+.card-section-with-title {
+  box-sizing: border-box;
+}
+
+.card-section-with-title_style_main {
+  padding: 100px 40px 0;
+}
+
+.card-section-with-title__title {
+  font-family: 'Didact-Gothic';
+  font-size: 40px;
+  font-weight: 400;
+  line-height: 1.3;
+  letter-spacing: -0.5px;
+  margin: 0;
+  /* padding-left: 40px; */
+  color: #21211b;
+}
+
+.card-section-with-title__header {
+  display: flex;
+  margin-bottom: 40px;
+}
