--- src/components/CatalogCardSection/CatalogCardSection.css
@@ -1,6 +1,0 @@
-.catalog-card-section {
-  margin: 0 auto;
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: 32px;
-}
