+++ src/components/PopularProducts/PopularProducts.css
@@ -1,3 +1,7 @@
 .popular-products {
+  display: grid;
+  grid-template-columns: repeat(4, auto);
+  gap: 32px;
+  justify-content: center;
   margin-bottom: 100px;
 }
